api_yamdb/requirements.txt
django-environ==0.8.1
asgiref==3.2.10
csvImporter==0.1.3.2
django==2.2.16
django-filter==2.4.0
django-adaptors==0.2.5
django-xml==3.0.0
django-import-export
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
Pillow==7.0.0
PyJWT==2.1.0
pytz==2020.1
requests==2.22.0
sqlparse==0.3.1